+++ src/alert.rs
@@ -34,7 +34,11 @@ use log::warn;
 
 fn main() {
     let arguments = alert_cli_parser::parse_arguments();
-    logging::initialise(arguments.occurrences_of(alert_cli_parser::FLAG_VERBOSE));
+    logging::initialise(
+        arguments
+            .value_of(alert_cli_parser::FLAG_LOG_CONFIG)
+            .expect("Missing default value in cli_parser"),
+    );
 
     debug!("Starting up");